+++ CHANGES.rst
@@ -38,6 +38,14 @@ Version 2.3.4
 
 *Unreleased*
 
+Improvements
+^^^^^^^^^^^^
+
+- Fail more gracefully is a user has an invalid locale set and fall back to the default
+  locale or English in case the default locale is invalid as well
+- Log an error if the configured default locale does not exist
+- Add ID-1 page size for badge printing (:pr:`4774`, thanks :user:`omegak`)
+
 Bugfixes
 ^^^^^^^^